agent/requirements.txt
faiss_cpu==1.9.0
langchain==0.3.10
langchain_community==0.3.10
langchain_core==0.3.22
langchain_openai==0.2.11
langchain_text_splitters==0.3.2
pandas==2.2.3
pydantic==2.10.3
python-dotenv==1.0.1
Requests==2.32.3

